added a universal subscriber/publisher mechanism for recording and logging data
