+++ app/src/main/java/com/dayani/m/roboplatform/CarManualControlActivity.java
@@ -14,18 +14,14 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.dayani.m.roboplatform.managers.CameraFlyVideo;
-import com.dayani.m.roboplatform.managers.MyLocationManager;
 import com.dayani.m.roboplatform.managers.MySensorManager;
 import com.dayani.m.roboplatform.managers.MyStateManager;
-import com.dayani.m.roboplatform.managers.MyStorageManager;
 import com.dayani.m.roboplatform.managers.MyUSBManager;
 import com.dayani.m.roboplatform.managers.MyWifiManager;
-import com.dayani.m.roboplatform.utils.ActivityRequirements;
 import com.dayani.m.roboplatform.utils.AppGlobals;
-import com.dayani.m.roboplatform.utils.ControlInput;
-import com.dayani.m.roboplatform.utils.SensorsContainer;
-import com.dayani.m.roboplatform.utils.ActivityRequirements.Requirement;
+import com.dayani.m.roboplatform.utils.data_types.ControlInput;
+import com.dayani.m.roboplatform.utils.data_types.SensorsContainer;
+import com.dayani.m.roboplatform.utils.interfaces.ActivityRequirements.Requirement;
 
 import java.util.ArrayList;
 import java.util.Arrays;
