+++ app/src/main/java/com/dayani/m/roboplatform/RecordSensorsActivity.java
@@ -17,31 +17,24 @@ import com.dayani.m.roboplatform.managers.MyBaseManager;
 import com.dayani.m.roboplatform.managers.MyLocationManager;
 import com.dayani.m.roboplatform.managers.MySensorManager;
 import com.dayani.m.roboplatform.managers.MyStorageManager;
-import com.dayani.m.roboplatform.utils.ActivityRequirements;
-import com.dayani.m.roboplatform.utils.SensorsViewModel;
+import com.dayani.m.roboplatform.utils.interfaces.ActivityRequirements;
+import com.dayani.m.roboplatform.utils.view_models.SensorsViewModel;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 
 /**
  * Responsible for all fragment interactions from requirement handling to recording
  */
 public class RecordSensorsActivity extends AppCompatActivity
-        implements MyStorageManager.StorageChannel,
-        ActivityRequirements.RequirementResolution {
+        implements ActivityRequirements.RequirementResolution {
 
     private static final String TAG = RecordSensorsActivity.class.getSimpleName();
 
-    public static final String EXTRA_KEY_RECORD_EXTERNAL = TAG+"key_record_external";
+    public static final String EXTRA_KEY_RECORD_EXTERNAL = TAG + "key_record_external";
 
-    boolean mbRecordExternal = false;
-
-    SensorsViewModel mVM_Sensors;
-
-    MyStorageManager mStorageManager;
-
-    private String mDsRoot;
+    private SensorsViewModel mVM_Sensors;
 
     /* -------------------------------------- Lifecycle ----------------------------------------- */
 
@@ -56,10 +49,9 @@ public class RecordSensorsActivity extends AppCompatActivity
 
         // resolve the record external sensor state (use USB device)
         Intent intent = getIntent();
-        mbRecordExternal = intent.getBooleanExtra(EXTRA_KEY_RECORD_EXTERNAL, false);
+        boolean mbRecordExternal = intent.getBooleanExtra(EXTRA_KEY_RECORD_EXTERNAL, false);
 
-        mDsRoot = initDsPath(mVM_Sensors);
-        initManagersAndSensors(mVM_Sensors);
+        initManagersAndSensors(mVM_Sensors, mbRecordExternal);
 
         // launch requirements fragment
         if (savedInstanceState == null) {
@@ -80,42 +72,51 @@ public class RecordSensorsActivity extends AppCompatActivity
         super.onDestroy();
     }
 
+    @Override
+    protected void onResume() {
+        super.onResume();
+
+        // update permissions state (in case user changes in the settings)
+        // TODO: find a broadcast or callback method to listen for these changes
+        for (MyBaseManager manager : mVM_Sensors.getAllManagers()) {
+            manager.updatePermissionsState(this);
+        }
+    }
+
     /* ------------------------------------ Init. & Clean --------------------------------------- */
 
-    private void initManagersAndSensors(SensorsViewModel vm) {
+    private void initManagersAndSensors(SensorsViewModel vm, boolean withUSB) {
 
-        mStorageManager = (MyStorageManager) getOrCreateManager(this, vm, MyStorageManager.class.getSimpleName());
+        MyStorageManager storageManager = (MyStorageManager) getOrCreateManager(
+                this, vm, MyStorageManager.class.getSimpleName());
 
-        getOrCreateManager(this, vm, MySensorManager.class.getSimpleName());
+        MyBaseManager manager = getOrCreateManager(this, vm, MySensorManager.class.getSimpleName());
+        manager.addChannelTransaction(storageManager);
 
-        getOrCreateManager(this, vm, MyLocationManager.class.getSimpleName());
+        manager = getOrCreateManager(this, vm, MyLocationManager.class.getSimpleName());
+        manager.addChannelTransaction(storageManager);
 
         // TODO: Add the rest [also change getOrCreateManager(...)]
+        //manager = getOrCreateManager(this, vm, CameraFlyVideo.class.getSimpleName());
+        //manager.addChannelTransaction(storageManager);
+
+        if (withUSB) {
+            Log.d(TAG, "USB manager is enabled");
+            //manager = getOrCreateManager(this, vm, MyUSBManager.class.getSimpleName());
+            //manager.addChannelTransaction(storageManager);
+        }
     }
 
     private void cleanManagers(List<MyBaseManager> lAllManagers) {
 
         for (MyBaseManager manager : lAllManagers) {
-            manager.clean();
+            manager.clean(this);
         }
     }
-
-    public static String initDsPath(SensorsViewModel vm) {
-
-        String curDsPath = vm.getDsPath().getValue();
-        if (curDsPath == null || curDsPath.isEmpty()) {
-
-            curDsPath = MyStorageManager.DS_FOLDER_PREFIX+MyStorageManager.getTimePrefix();
-            vm.setDsPath(curDsPath);
-            Log.v(TAG, "Created new ds root: "+curDsPath);
-        }
-        return curDsPath;
-    }
-
-    /* --------------------------------------- Helpers ------------------------------------------ */
 
     // don't repeat the same check and create in all fragments and activities
-    public static MyBaseManager getOrCreateManager(Context context, SensorsViewModel vm, String managerClassName) {
+    public static MyBaseManager getOrCreateManager(Context context, SensorsViewModel vm,
+                                                   String managerClassName) {
 
         MyBaseManager manager = null;
 
@@ -140,107 +141,51 @@ public class RecordSensorsActivity extends AppCompatActivity
                 vm.addManagerAndSensors(context, manager);
             }
         }
+        //TODO: Add the rest
 
         return manager;
     }
 
-    /* --------------------------------- Storage Transactions ----------------------------------- */
-
-    // Better to define this interface in the storage creator context
-    @Override
-    public int getStorageChannel(List<String> folders, String fileName, boolean append) {
-
-        if (mStorageManager != null) {
-
-            List<String> foldersWithRoot = new ArrayList<>();
-
-            if (mDsRoot == null) {
-                mDsRoot = initDsPath(mVM_Sensors);
-            }
-            foldersWithRoot.add(mDsRoot);
-            foldersWithRoot.addAll(folders);
-
-            String[] foldersArr = foldersWithRoot.toArray(new String[0]);
-
-            String fullPath = mStorageManager.resolveFilePath(this, foldersArr, fileName);
-            return mStorageManager.subscribeStorageChannel(this, fullPath, append);
-        }
-        else {
-            Log.w(TAG, "Storage manager is empty");
-        }
-        return -1;
-    }
-
-    @Override
-    public void publishMessage(int id, String msg) {
-
-        if (mStorageManager != null) {
-            mStorageManager.publishMessage(this, id, msg);
-        }
-        else {
-            Log.w(TAG, "Storage manager is empty");
-        }
-    }
-
-    @Override
-    public void resetChannel(int id) {
-
-        if (mStorageManager != null) {
-            mStorageManager.resetStorageChannel(this, id);
-        }
-        else {
-            Log.w(TAG, "Storage manager is empty");
-        }
-    }
-
-    @Override
-    public String getFullFilePath(int id) {
-
-        if (mStorageManager != null) {
-            return mStorageManager.getFullFilePath(this, id);
-        }
-        else {
-            Log.w(TAG, "Storage manager is empty");
-            return "";
-        }
-    }
-
-    @Override
-    public void removeChannel(int id) {
-
-        if (mStorageManager != null) {
-            mStorageManager.removeChannel(this, id);
-        }
-        else {
-            Log.w(TAG, "Storage manager is empty");
-        }
-    }
-
     /* ---------------------------------- Request Resolutions ----------------------------------- */
-
-    private final ActivityResultLauncher<Intent> mIntentLauncher = registerForActivityResult(
-            new ActivityResultContracts.StartActivityForResult(),
-            this::processActivityResult);
-
-    private final ActivityResultLauncher<IntentSenderRequest> mIntentSender = registerForActivityResult(
-            new ActivityResultContracts.StartIntentSenderForResult(),
-            this::processActivityResult);
 
     private final ActivityResultLauncher<String[]> mPermsLauncher = registerForActivityResult(
             new ActivityResultContracts.RequestMultiplePermissions(),
-            permissions -> {
-                if (mVM_Sensors != null && permissions != null) {
-                    for (MyBaseManager manager : mVM_Sensors.getAllManagers()) {
-                        manager.onPermissionsResult(this, permissions);
-                    }
-                }
+            this::processPermissions);
+
+    private int mRequestCode = 0;
+
+    private final ActivityResultLauncher<Intent> mIntentLauncher = registerForActivityResult(
+            new ActivityResultContracts.StartActivityForResult(),
+            result -> {
+                this.processActivityResult(mRequestCode, result);
             });
 
-    private void processActivityResult(ActivityResult result) {
+    private final ActivityResultLauncher<IntentSenderRequest> mIntentSender = registerForActivityResult(
+            new ActivityResultContracts.StartIntentSenderForResult(),
+            result -> {
+                this.processActivityResult(mRequestCode, result);
+            });
 
-        if (mVM_Sensors != null && result != null) {
+    private void processActivityResult(int requestCode, ActivityResult result) {
+
+        Intent data = result.getData();
+        if (data == null) {
+            return;
+        }
+        data.putExtra(MyBaseManager.KEY_INTENT_ACTIVITY_LAUNCHER, requestCode);
+
+        if (mVM_Sensors != null) {
             for (MyBaseManager manager : mVM_Sensors.getAllManagers()) {
                 manager.onActivityResult(this, result);
+            }
+        }
+    }
+
+    private void processPermissions(Map<String, Boolean> permissions) {
+
+        if (mVM_Sensors != null && permissions != null) {
+            for (MyBaseManager manager : mVM_Sensors.getAllManagers()) {
+                manager.onPermissionsResult(this, permissions);
             }
         }
     }
@@ -252,14 +197,22 @@ public class RecordSensorsActivity extends AppCompatActivity
     }
 
     @Override
-    public void requestResolution(Intent activityIntent) {
+    public void requestResolution(int requestCode, Intent activityIntent) {
 
+        if (activityIntent == null) {
+            return;
+        }
+        mRequestCode = requestCode;
         mIntentLauncher.launch(activityIntent);
     }
 
     @Override
-    public void requestResolution(IntentSenderRequest resolutionIntent) {
+    public void requestResolution(int requestCode, IntentSenderRequest resolutionIntent) {
 
+        if (resolutionIntent == null) {
+            return;
+        }
+        mRequestCode = requestCode;
         mIntentSender.launch(resolutionIntent);
     }
