+++ app/src/main/java/com/dayani/m/roboplatform/dump/RecordSensorsActivity_old2.java
@@ -38,10 +38,9 @@ import com.dayani.m.roboplatform.managers.MySensorManager;
 import com.dayani.m.roboplatform.managers.MyStateManager;
 import com.dayani.m.roboplatform.managers.MyStorageManager;
 import com.dayani.m.roboplatform.managers.MyUSBManager;
-import com.dayani.m.roboplatform.utils.ActivityRequirements;
 import com.dayani.m.roboplatform.utils.AppGlobals;
-import com.dayani.m.roboplatform.utils.AutoFitTextureView;
-import com.dayani.m.roboplatform.utils.SensorsContainer;
+import com.dayani.m.roboplatform.utils.cutom_views.AutoFitTextureView;
+import com.dayani.m.roboplatform.utils.data_types.SensorsContainer;
 
 import java.io.File;
