+++ app/src/main/java/com/dayani/m/roboplatform/requirements/LocationReqFragment.java
@@ -16,8 +16,8 @@ import android.view.ViewGroup;
 import com.dayani.m.roboplatform.R;
 import com.dayani.m.roboplatform.RequirementsFragment;
 import com.dayani.m.roboplatform.managers.MyLocationManager;
-import com.dayani.m.roboplatform.utils.ActivityRequirements.Requirement;
-import com.dayani.m.roboplatform.utils.SensorsViewModel;
+import com.dayani.m.roboplatform.utils.interfaces.ActivityRequirements.Requirement;
+import com.dayani.m.roboplatform.utils.view_models.SensorsViewModel;
 
 import java.util.List;
