--- app/src/main/java/com/dayani/m/roboplatform/utils/StrBufferLogger.java
@@ -1,4 +1,0 @@
-package com.dayani.m.roboplatform.utils;
-
-public class StrBufferLogger {
-}
