+++ app/src/main/java/com/dayani/m/roboplatform/utils/helpers/MyScreenOperations.java
@@ -1,4 +1,4 @@
-package com.dayani.m.roboplatform.utils;
+package com.dayani.m.roboplatform.utils.helpers;
 
 import android.app.Activity;
 import android.content.pm.ActivityInfo;
