+++ app/src/main/java/com/dayani/m/roboplatform/utils/interfaces/MyFragmentInteraction.java
@@ -1,4 +1,4 @@
-package com.dayani.m.roboplatform.utils;
+package com.dayani.m.roboplatform.utils.interfaces;
 
 import androidx.fragment.app.Fragment;
