+++ app/src/main/java/com/dayani/m/roboplatform/utils/loggers/StrBufferLogger.java
@@ -1,0 +1,4 @@
+package com.dayani.m.roboplatform.utils.loggers;
+
+public class StrBufferLogger {
+}
