+++ app/src/main/res/layout/activity_record_all.xml
@@ -37,7 +37,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.dayani.m.roboplatform.utils.AutoFitTextureView
+    <com.dayani.m.roboplatform.utils.cutom_views.AutoFitTextureView
         android:id="@+id/texture"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
